samsung ics opensource update 1

+++ mm/oom_kill.c
@@ -162,7 +162,7 @@ static bool oom_unkillable_task(struct task_struct *p,
 unsigned int oom_badness(struct task_struct *p, struct mem_cgroup *mem,
 		      const nodemask_t *nodemask, unsigned long totalpages)
 {
-	long points;
+	int points;
 
 	if (oom_unkillable_task(p, mem, nodemask))
 		return 0;
